249 : Moved Wrapper to its own file.
